Création classe abstraite Gestion et recuperation alerte panneauSolaire
