+++ app/IOT/config.ini
@@ -4,7 +4,7 @@ port = 1883
 topic_subscribe = solaredge/#,AM107/by-room/
 
 [DATA]
-salles = C102,E007
+salles = 
 types = temperature,humidity,co2,room
 typesSolaire = lifeTimeData,lastUpdateTime,currentPower
 frequency = 120
@@ -18,5 +18,5 @@ alert_file = alertJSON.json
 seuil_temperature = 25
 seuil_co2 = 800
 seuil_humidity = 60
-seuil_power = 3000
-seuil_energy = 10000
+seuil_power = 3
+seuil_energy = 1
