+++ app/IOT/src/main/java/Data/GestionData.java
@@ -4,26 +4,17 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.List;
 import java.util.Map;
 
-public class GestionData {
-
+public class GestionData extends Gestion {
     @JsonProperty("Capteurs AM")
     private Map<String, List<Capteur>> capteursAM;
 
-    @JsonProperty("PanneauSolaire")
-    private List<PanneauSolaire> donneesPanneau;
-
-    // Getters et Setters pour capteursAM et panneaux
-    public Map<String, List<Capteur>> getCapteursAM() {
-        return capteursAM;
-    }
     public void setCapteursAM(Map<String, List<Capteur>> capteursAM) {
         this.capteursAM = capteursAM;
     }
-    public List<PanneauSolaire> getPanneaux() {
-        return donneesPanneau;
-    }
-    public void setPanneaux(List<PanneauSolaire> pdonneesPanneau) {
-        this.donneesPanneau = pdonneesPanneau;
+    
+    @Override
+    public Map<String, List<Capteur>> getCapteurs() {
+        return capteursAM;
     }
 
 }
